adjusted locations of functions
