+++ brain_observatory_utilities/datasets/behavior/data_formatting.py
@@ -1,0 +1,192 @@
+import pandas as pd
+from allensdk.brain_observatory.behavior.trials_processing import calculate_reward_rate
+from brain_observatory_utilities.utilities.general_utilities import get_trace_average
+
+
+
+def add_mean_running_speed_to_stimulus_presentations(stimulus_presentations,
+                                                     running_speed,
+                                                     time_window=[0, 0.75]):
+    '''
+    Append a column to stimulus_presentations which contains
+    the mean running speed in a range relative to
+    the stimulus start time.
+
+    Args:
+        stimulus_presentations (pd.DataFrame): dataframe of
+            stimulus presentations.
+            Must contain: 'start_time'
+        running_speed (pd.DataFrame): dataframe of running speed.
+            Must contain: 'speed', 'timestamps'
+        time_window: array
+            timestamps in seconds, relative to the start of each stimulus
+            to average the running speed.
+            default = [-3,3]
+    Returns:
+        stimulus_presentations with new column
+        "mean_running_speed" containing the
+        mean running speed within the specified window
+        following each stimulus presentation.
+
+    Example:
+        # get visual behavior cache
+        from allensdk.brain_observatory.behavior.behavior_project_cache import VisualBehaviorOphysProjectCache as bpc  # noqa E501
+        cache_dir = SOME_LOCAL_DIR
+        cache = bpc.from_s3_cache(cache_dir=cache_dir)
+
+        # load data for one experiment
+        ophys_experiment = cache.get_behavior_ophys_experiment(experiment_id)
+
+        # get necessary tables
+        stimulus_presentations = ophys_experiment.stimulus_presentations.copy()
+        running_speed = ophys_experiment.running_speed.copy()
+
+        # add running_speed to stim presentations
+        stimulus_presentations = add_mean_running_speed_to_stimulus_presentations(stimulus_presentations, running_speed)  # noqa E501
+    '''
+
+    stim_running_speed = stimulus_presentations.apply(
+        lambda row: get_trace_average(
+            running_speed['speed'].values,
+            running_speed['timestamps'].values,
+            row["start_time"] + time_window[0],
+            row["start_time"] + time_window[1]), axis=1,)
+    stimulus_presentations["mean_running_speed"] = stim_running_speed
+    return stimulus_presentations
+
+
+def add_rewards_to_stimulus_presentations(stimulus_presentations,
+                                          rewards,
+                                          time_window=[0,3]):
+    '''
+    Append a column to stimulus_presentations which contains
+    the timestamps of rewards that occur
+    in a range relative to the onset of the stimulus.
+
+    Args:
+        stimulus_presentations (pd.DataFrame): dataframe of
+            stimulus presentations.
+            Must contain: 'start_time'
+        rewards (pd.DataFrame): rewards dataframe. Must contain 'timestamps'
+        time_window (list with 2 elements): start and end of the range
+            relative to the start of each stimulus
+            to average the running speed.
+    Returns:
+        stimulus_presentations with a new column called "reward" that contains
+        reward times that fell within the window relative to each stim time
+
+    Example:
+        # get visual behavior cache
+        from allensdk.brain_observatory.behavior.behavior_project_cache import VisualBehaviorOphysProjectCache as bpc  # noqa E501
+        cache_dir = SOME_LOCAL_DIRECTORY
+        cache = bpc.from_s3_cache(cache_dir=cache_dir)
+
+        # load data for one experiment
+        ophys_experiment = cache.get_behavior_ophys_experiment(experiment_id)
+
+        # get necessary tables
+        stimulus_presentations = ophys_experiment.stimulus_presentations.copy()
+        rewards = ophys_experiment.rewards.copy()
+
+        # add rewards to stim presentations
+        stimulus_presentations = add_rewards_to_stimulus_presentations(stimulus_presentations, rewards)  # noqa E501
+    '''
+
+    reward_times = rewards['timestamps'].values
+    rewards_each_stim = stimulus_presentations.apply(
+        lambda row: reward_times[
+            ((reward_times > row["start_time"] + time_window[0]) & (reward_times < row["start_time"] + time_window[1]))],
+        axis=1,
+    )
+    stimulus_presentations["rewards"] = rewards_each_stim
+    return stimulus_presentations
+
+
+def add_licks_to_stimulus_presentations(stimulus_presentations,
+                                        licks,
+                                        time_window=[0, 0.75]):
+    '''
+    Append a column to stimulus_presentations which
+    contains the timestamps of licks that occur
+    in a range relative to the onset of the stimulus.
+
+    Args:
+        stimulus_presentations (pd.DataFrame): 
+            dataframe of stimulus presentations.
+            Must contain: 'start_time'
+        licks (pd.DataFrame): lick dataframe. Must contain 'timestamps'
+        time_window (list with 2 elements): start and end of the range
+            relative to the start of each stimulus to average the running speed.  # noqa E501
+    Returns:
+        stimulus_presentations with a new column called "licks" that contains
+        lick times that fell within the window relative to each stim time
+
+
+    Example:
+        # get visual behavior cache
+        from allensdk.brain_observatory.behavior.behavior_project_cache import VisualBehaviorOphysProjectCache as bpc  # noqa E501
+        cache_dir = SOME_LOCAL_DIRECTORY
+        cache = bpc.from_s3_cache(cache_dir=cache_dir)
+
+        # load data for one experiment
+        ophys_experiment = cache.get_behavior_ophys_experiment(experiment_id)
+
+        # get necessary tables
+        stimulus_presentations = ophys_experiment.stimulus_presentations.copy()
+        licks = ophys_experiment.licks.copy()
+
+        # add licks to stim presentations
+        stimulus_presentations = add_licks_to_stimulus_presentations(stimulus_presentations, licks)
+    '''
+
+    lick_times = licks['timestamps'].values
+    licks_each_stim = stimulus_presentations.apply(
+        lambda row: lick_times[
+            ((lick_times > row["start_time"] + time_window[0]) & (lick_times < row["start_time"] + time_window[1]))],
+        axis=1,
+    )
+    stimulus_presentations["licks"] = licks_each_stim
+    return stimulus_presentations
+
+
+def add_reward_rate_to_stimulus_presentations(stimulus_presentations,
+                                              trials):
+    '''
+    Parameters:
+    ____________
+    trials: Pandas.DataFrame
+        ophys_experiment.trials
+    stimulus_presentation: Pandas.DataFrame
+        ophys_experiment.stimulus_presentations
+
+    Returns:
+    ___________
+    stimulus_presentation: Pandas.DataFrame
+        with 'reward_rate_trials' column
+
+    'reward_rate_trials' is calculated by the SDK based on the rolling reward rate over trials (not stimulus presentations)
+    https://github.com/AllenInstitute/AllenSDK/blob/master/allensdk/brain_observatory/behavior/trials_processing.py#L941
+    '''
+
+    last_time = 0
+    reward_rate_by_frame = []
+    if 'reward_rate' not in trials:
+        trials['reward_rate'] = calculate_reward_rate(trials['response_latency'].values,
+                                                      trials['start_time'],
+                                                      window=.5)
+
+    trials = trials[trials['aborted'] == False]  # NOQA
+    for change_time in trials.change_time.values:
+        reward_rate = trials[trials.change_time ==  # NOQA
+                             change_time].reward_rate.values[0]
+        for start_time in stimulus_presentations.start_time:
+            if (start_time < change_time) and (start_time > last_time):
+                reward_rate_by_frame.append(reward_rate)
+                last_time = start_time
+    # fill the last flashes with last value
+    for i in range(len(stimulus_presentations) - len(reward_rate_by_frame)):
+        reward_rate_by_frame.append(reward_rate_by_frame[-1])
+
+    stimulus_presentations['reward_rate_trials'] = reward_rate_by_frame
+    return stimulus_presentations
+
