+++ brain_observatory_utilities/datasets/optical_physiology/data_formatting.py
@@ -1,8 +1,5 @@
 import pandas as pd
 import numpy as np
-import xarray as xr
-from tqdm import tqdm
-from allensdk.brain_observatory.behavior.trials_processing import calculate_reward_rate
 
 
 def build_tidy_cell_df(ophys_experiment, exclude_invalid_rois=True):
@@ -115,251 +112,6 @@ def get_event_timestamps(
 
     return event_times, event_ids
 
-
-def add_rewards_to_stimulus_presentations(
-    stimulus_presentations,
-    rewards,
-    time_window=[
-        0,
-        3]):
-    '''
-    Append a column to stimulus_presentations which contains
-    the timestamps of rewards that occur
-    in a range relative to the onset of the stimulus.
-
-    Args:
-        stimulus_presentations (pd.DataFrame): dataframe of
-            stimulus presentations.
-            Must contain: 'start_time'
-        rewards (pd.DataFrame): rewards dataframe. Must contain 'timestamps'
-        time_window (list with 2 elements): start and end of the range
-            relative to the start of each stimulus
-            to average the running speed.
-    Returns:
-        stimulus_presentations with a new column called "reward" that contains
-        reward times that fell within the window relative to each stim time
-
-    Example:
-        # get visual behavior cache
-        from allensdk.brain_observatory.behavior.behavior_project_cache import VisualBehaviorOphysProjectCache as bpc  # noqa E501
-        cache_dir = SOME_LOCAL_DIRECTORY
-        cache = bpc.from_s3_cache(cache_dir=cache_dir)
-
-        # load data for one experiment
-        ophys_experiment = cache.get_behavior_ophys_experiment(experiment_id)
-
-        # get necessary tables
-        stimulus_presentations = ophys_experiment.stimulus_presentations.copy()
-        rewards = ophys_experiment.rewards.copy()
-
-        # add rewards to stim presentations
-        stimulus_presentations = add_rewards_to_stimulus_presentations(stimulus_presentations, rewards)  # noqa E501
-    '''
-
-    reward_times = rewards['timestamps'].values
-    rewards_each_stim = stimulus_presentations.apply(
-        lambda row: reward_times[
-            ((reward_times > row["start_time"] + time_window[0]) & (reward_times < row["start_time"] + time_window[1]))],
-        axis=1,
-    )
-    stimulus_presentations["rewards"] = rewards_each_stim
-    return stimulus_presentations
-
-
-def add_licks_to_stimulus_presentations(stimulus_presentations, licks, time_window=[0, 0.75]):
-    '''
-    Append a column to stimulus_presentations which
-    contains the timestamps of licks that occur
-    in a range relative to the onset of the stimulus.
-
-    Args:
-        stimulus_presentations (pd.DataFrame): 
-            dataframe of stimulus presentations.
-            Must contain: 'start_time'
-        licks (pd.DataFrame): lick dataframe. Must contain 'timestamps'
-        time_window (list with 2 elements): start and end of the range
-            relative to the start of each stimulus to average the running speed.  # noqa E501
-    Returns:
-        stimulus_presentations with a new column called "licks" that contains
-        lick times that fell within the window relative to each stim time
-
-
-    Example:
-        # get visual behavior cache
-        from allensdk.brain_observatory.behavior.behavior_project_cache import VisualBehaviorOphysProjectCache as bpc  # noqa E501
-        cache_dir = SOME_LOCAL_DIRECTORY
-        cache = bpc.from_s3_cache(cache_dir=cache_dir)
-
-        # load data for one experiment
-        ophys_experiment = cache.get_behavior_ophys_experiment(experiment_id)
-
-        # get necessary tables
-        stimulus_presentations = ophys_experiment.stimulus_presentations.copy()
-        licks = ophys_experiment.licks.copy()
-
-        # add licks to stim presentations
-        stimulus_presentations = add_licks_to_stimulus_presentations(stimulus_presentations, licks)
-    '''
-
-    lick_times = licks['timestamps'].values
-    licks_each_stim = stimulus_presentations.apply(
-        lambda row: lick_times[
-            ((lick_times > row["start_time"] + time_window[0]) & (lick_times < row["start_time"] + time_window[1]))],
-        axis=1,
-    )
-    stimulus_presentations["licks"] = licks_each_stim
-    return stimulus_presentations
-
-
-def get_trace_average(trace, timestamps, start_time, stop_time):
-    """
-    takes average value of a trace within a window
-    designated by start_time and stop_time
-    """
-    values_this_range = trace[(
-        (timestamps >= start_time) & (timestamps < stop_time))]
-    return values_this_range.mean()
-
-
-def add_mean_running_speed_to_stimulus_presentations(stimulus_presentations, running_speed, time_window=[0, 0.75]):
-    '''
-    Append a column to stimulus_presentations which contains
-    the mean running speed in a range relative to
-    the stimulus start time.
-
-    Args:
-        stimulus_presentations (pd.DataFrame): dataframe of
-            stimulus presentations.
-            Must contain: 'start_time'
-        running_speed (pd.DataFrame): dataframe of running speed.
-            Must contain: 'speed', 'timestamps'
-        time_window: array
-            timestamps in seconds, relative to the start of each stimulus
-            to average the running speed.
-            default = [-3,3]
-    Returns:
-        stimulus_presentations with new column
-        "mean_running_speed" containing the
-        mean running speed within the specified window
-        following each stimulus presentation.
-
-    Example:
-        # get visual behavior cache
-        from allensdk.brain_observatory.behavior.behavior_project_cache import VisualBehaviorOphysProjectCache as bpc  # noqa E501
-        cache_dir = SOME_LOCAL_DIR
-        cache = bpc.from_s3_cache(cache_dir=cache_dir)
-
-        # load data for one experiment
-        ophys_experiment = cache.get_behavior_ophys_experiment(experiment_id)
-
-        # get necessary tables
-        stimulus_presentations = ophys_experiment.stimulus_presentations.copy()
-        running_speed = ophys_experiment.running_speed.copy()
-
-        # add running_speed to stim presentations
-        stimulus_presentations = add_mean_running_speed_to_stimulus_presentations(stimulus_presentations, running_speed)  # noqa E501
-    '''
-
-    stim_running_speed = stimulus_presentations.apply(
-        lambda row: get_trace_average(
-            running_speed['speed'].values,
-            running_speed['timestamps'].values,
-            row["start_time"] + time_window[0],
-            row["start_time"] + time_window[1]), axis=1,)
-    stimulus_presentations["mean_running_speed"] = stim_running_speed
-    return stimulus_presentations
-
-
-def add_mean_pupil_to_stimulus_presentations(stimulus_presentations, eye_tracking, column_to_use='pupil_area', time_window=[0, 0.75]):
-    '''
-    Append a column to stimulus_presentations which contains
-    the mean pupil area, diameter, or radius in a range relative to
-    the stimulus start time.
-
-    Args:
-        stimulus_presentations (pd.DataFrame): dataframe of stimulus presentations.  # noqa E501
-            Must contain: 'start_time'
-        eye_tracking (pd.DataFrame): dataframe of eye tracking data.
-            Must contain: timestamps', 'pupil_area', 'pupil_width', 'likely_blinks'
-        time_window (list with 2 elements): start and end of the range
-            relative to the start of each stimulus to average the pupil area.
-        column_to_use: column in eyetracking table to use to get mean, options: 'pupil_area', 'pupil_width', 'pupil_radius', 'pupil_diameter'
-                        if 'pupil_diameter' or 'pupil_radius' are provided, they will be calculated from 'pupil_area'
-                        if 'pupil_width' is provided, the column 'pupil_width' will be directly used from eye_tracking table
-    Returns:
-        stimulus_presentations table with new column "mean_pupil_"+column_to_use with the
-        mean pupil value within the specified window following each stimulus presentation.
-
-    Example:
-        # get visual behavior cache
-        from allensdk.brain_observatory.behavior.behavior_project_cache import VisualBehaviorOphysProjectCache as bpc  # noqa E501
-        cache_dir = r'\\allen\programs\braintv\workgroups\nc-ophys\visual_behavior\platform_paper_cache'
-        cache = bpc.from_s3_cache(cache_dir=cache_dir)
-
-        # load data for one experiment
-        ophys_experiment = cache.get_behavior_ophys_experiment(experiment_id)
-
-        # get necessary tables
-        stimulus_presentations = ophys_experiment.stimulus_presentations.copy()
-        eye_tracking = ophys_experiment.eye_tracking.copy()
-
-        # add pupil area to stim presentations
-        stimulus_presentations = add_mean_pupil_to_stimulus_presentations(stimulus_presentations, eye_tracking, column_to_use='pupil_area')  # noqa E501
-    '''
-
-    eye_tracking = get_pupil_data(eye_tracking, interpolate_likely_blinks=True, normalize_to_gray_screen=True, zscore=False,
-                   interpolate_to_ophys=False, ophys_timestamps=None, stimulus_presentations=stimulus_presentations)
-
-    eye_tracking_timeseries = eye_tracking[column_to_use].values
-    mean_pupil_around_stimulus = stimulus_presentations.apply(
-        lambda row: get_trace_average(
-            eye_tracking_timeseries,
-            eye_tracking['timestamps'].values,
-            row["start_time"] + time_window[0],
-            row["start_time"] + time_window[1],
-        ), axis=1,)
-    stimulus_presentations["mean_"+column_to_use] = mean_pupil_around_stimulus
-    return stimulus_presentations
-
-
-def add_reward_rate_to_stimulus_presentations(stimulus_presentations, trials):
-    '''
-    Parameters:
-    ____________
-    trials: Pandas.DataFrame
-        ophys_experiment.trials
-    stimulus_presentation: Pandas.DataFrame
-        ophys_experiment.stimulus_presentations
-
-    Returns:
-    ___________
-    stimulus_presentation: Pandas.DataFrame
-        with 'reward_rate_trials' column
-
-    'reward_rate_trials' is calculated by the SDK based on the rolling reward rate over trials (not stimulus presentations)
-    https://github.com/AllenInstitute/AllenSDK/blob/master/allensdk/brain_observatory/behavior/trials_processing.py#L941
-    '''
-
-    last_time = 0
-    reward_rate_by_frame = []
-    if 'reward_rate' not in trials:
-        trials['reward_rate'] = calculate_reward_rate(trials['response_latency'].values,
-                                                  trials['start_time'], window=.5)
-
-    trials = trials[trials['aborted'] == False]  # NOQA
-    for change_time in trials.change_time.values:
-        reward_rate = trials[trials.change_time ==  # NOQA
-                             change_time].reward_rate.values[0]
-        for start_time in stimulus_presentations.start_time:
-            if (start_time < change_time) and (start_time > last_time):
-                reward_rate_by_frame.append(reward_rate)
-                last_time = start_time
-    # fill the last flashes with last value
-    for i in range(len(stimulus_presentations) - len(reward_rate_by_frame)):
-        reward_rate_by_frame.append(reward_rate_by_frame[-1])
-
-    stimulus_presentations['reward_rate_trials'] = reward_rate_by_frame
-    return stimulus_presentations
 
 
 def add_epochs_to_stimulus_presentations(stimulus_presentations, time_column='start_time', epoch_duration_mins=10):
@@ -488,52 +240,6 @@ def add_time_from_last_change_to_stimulus_presentations(stimulus_presentations):
     stimulus_presentations["time_from_last_change"] = time_from_last_change
 
     return stimulus_presentations
-
-
-
-def get_annotated_stimulus_presentations(ophys_experiment):
-    """
-    Takes in an ophys_experiment dataset object and returns the stimulus_presentations table with additional columns.
-    Adds several useful columns to the stimulus_presentations table, including the mean running speed and pupil diameter for each stimulus,
-    the times of licks for each stimulus, the rolling reward rate, an identifier for 10 minute epochs within a session,
-    whether or not a stimulus was a pre-change or pre or post omission, and whether change stimuli were hits or misses
-    :param ophys_experiment: obj
-        AllenSDK BehaviorOphysExperiment object
-        A BehaviorOphysExperiment instance
-        See https://github.com/AllenInstitute/AllenSDK/blob/master/allensdk/brain_observatory/behavior/behavior_ophys_ophys_experiment.py  # noqa E501
-    :return: stimulus_presentations attribute of BehaviorOphysExperiment, with additional columns added
-    """
-    stimulus_presentations = ophys_experiment.stimulus_presentations.copy()
-    stimulus_presentations = add_licks_to_stimulus_presentations(stimulus_presentations, ophys_experiment.licks,
-                                                                  time_window=[0, 0.75])
-    stimulus_presentations = add_mean_running_speed_to_stimulus_presentations(stimulus_presentations,
-                                                                               ophys_experiment.running_speed,
-                                                                               time_window=[0, 0.75])
-    try:
-        stimulus_presentations = add_mean_pupil_to_stimulus_presentations(stimulus_presentations,
-                                                                           ophys_experiment.eye_tracking,
-                                                                           column_to_use='pupil_width',
-                                                                           time_window=[0, 0.75])
-    except:
-        print('could not add mean pupil to stimulus presentations, length of eye_tracking attribute is', len(ophys_experiment.eye_tracking))
-    stimulus_presentations = add_reward_rate_to_stimulus_presentations(stimulus_presentations, ophys_experiment.trials)
-    stimulus_presentations = add_epochs_to_stimulus_presentations(stimulus_presentations,
-                                                                   time_column='start_time',
-                                                                   epoch_duration_mins=10)
-    stimulus_presentations = add_trials_data_to_stimulus_presentations_table(stimulus_presentations, ophys_experiment.trials)
-    # add engagement state based on reward rate - note this reward rate is calculated differently than the SDK version
-    stimulus_presentations = add_engagement_state_to_stimulus_presentations(stimulus_presentations, ophys_experiment.trials)
-    # add time from last change
-    stimulus_presentations = add_time_from_last_change_to_stimulus_presentations(stimulus_presentations)
-    # add pre-change, pre-omission, lick on next flash etc
-    stimulus_presentations['pre_change'] = stimulus_presentations['is_change'].shift(-1)
-    stimulus_presentations['pre_omitted'] = stimulus_presentations['omitted'].shift(-1)
-    stimulus_presentations['post_omitted'] = stimulus_presentations['omitted'].shift(1)
-    stimulus_presentations['licked'] = [True if len(licks) > 0 else False for licks in stimulus_presentations.licks.values]
-    stimulus_presentations['lick_on_next_flash'] = stimulus_presentations['licked'].shift(-1)
-
-    return stimulus_presentations
-
 
 
 def annotate_stimuli(dataset, inplace=False):
@@ -711,32 +417,5 @@ def calculate_response_matrix(stimuli, aggfunc=np.mean, sort_by_column=True, eng
     return response_matrix
 
 
-def get_licks_df(ophys_experiment):
-    '''
-    Creates a dataframe containing columns for 'timestamps', 'licks', where values are from
-    a binary array of the length of stimulus timestamps where frames with no lick are 0 and frames with a lick are 1,
-    and a column called 'lick_rate' with values of 'licks' averaged over a 6 frame window to get licks per 100ms,
-    Can be used to plot event triggered average lick rate
-    Parameters:
-    -----------
-    ophys_experiment: obj
-        AllenSDK BehaviorOphysExperiment object
-        A BehaviorOphysExperiment instance
-        See https://github.com/AllenInstitute/AllenSDK/blob/master/allensdk/brain_observatory/behavior/behavior_ophys_ophys_experiment.py  # noqa E501
-
-    Returns:
-    --------
-    Pandas.DataFrame with columns 'timestamps', 'licks', and 'lick_rate' in units of licks / 100ms
-
-    '''
-    timestamps = ophys_experiment.stimulus_timestamps.copy()
-    licks = ophys_experiment.licks.copy()
-    lick_array = np.zeros(timestamps.shape)
-    lick_array[licks.frame.values] = 1
-    licks_df = pd.DataFrame(data=timestamps, columns=['timestamps'])
-    licks_df['licks'] = lick_array
-    licks_df['lick_rate'] = licks_df['licks'].rolling(window=6, min_periods=1, win_type='triang').mean()
-
-    return licks_df
