+++ brain_observatory_utilities/utilities/general_utilities.py
@@ -1,0 +1,11 @@
+
+
+
+def get_trace_average(trace, timestamps, start_time, stop_time):
+    """
+    takes average value of a trace within a window
+    designated by start_time and stop_time
+    """
+    values_this_range = trace[(
+        (timestamps >= start_time) & (timestamps < stop_time))]
+    return values_this_range.mean()
